+++ normtest/__init__.py
@@ -1,9 +1,8 @@
 # https://stackoverflow.com/a/16084844/17872198
-from . version import __author__
-from . version import __email__
-from . version import __name__
-from . version import __version__
+from .version import __author__
+from .version import __email__
+from .version import __name__
+from .version import __version__
 
 
-
-from .normtest import ryan_joiner
+# from .normtest import ryan_joiner