+++ normtest/ryan_joiner.py
@@ -1,9 +1,9 @@
-"""
+"""This module contains functions related to the Ryan-Joiner test
 
 ##### List of functions (alphabetical order) #####
 
 ## Functions WITH good TESTS ###
-
+- order_statistic(sample_size, alpha="3/8", safe=False)
 
 ## Functions WITH some TESTS ###
 
@@ -21,7 +21,7 @@
 - make_bar_plot(axes, df, n_samples, alpha_column_name=None, n_rep_name=None, tests_column_names=None, normal=True, safe=False)
 - make_heatmap(axes, df, n_samples, alpha_column_name=None, n_rep_name=None, tests_column_names=None, normal=True, safe=False)
 - normal_distribution_plot(axes, n_rep, seed=None, xinfo=[0.00001, 0.99999, 1000], loc=0.0, scale=1.0, safe=False)
-- ordered_statistics(n, method)
+
 
 
 ##### List of CLASS (alphabetical order) #####
@@ -32,9 +32,9 @@
 
 Author: Anderson Marcos Dias Canteli <andersonmdcanteli@gmail.com>
 
-Created: September 22, 2023.
+Last update: November 02, 2023
 
-Last update: October 04, 2023
+Last update: November 02, 2023
 
 
 
@@ -51,11 +51,13 @@ import matplotlib.pyplot as plt
 import pandas as pd
 from scipy import stats
 from scipy import interpolate
-import seaborn as sns
 
-### home made ###
-from .utils import checkers
-from .utils import constants
+# import seaborn as sns
+
+### self made ###
+from paramcheckup import parameters, types, numbers
+
+# from .utils import constants
 
 ##### CONSTANTS #####
 
@@ -65,4 +67,87 @@ from .utils import constants
 ##### FUNCTIONS #####
 
 
-## RYAN - JOINER TEST ##
+def order_statistic(sample_size, alpha="3/8", safe=False):
+    """This function estimates the normal statistical order (:math:`p_{i}`) using approximations [1]_.
+
+    Parameters
+    ----------
+    sample_size : int
+        The sample size. Must be equal or greater than `4`;
+    alpha : str, optional
+        A `str` with the alpha value that should be adopted (See details in the Notes section). The options are:
+         * `"0"`;
+         * `"3/8"` (default);
+         * `"1/2"`;
+
+    safe : bool, optional
+        Whether to check the inputs before performing the calculations (`True`) or not (`False`, default). Useful for beginners to identify problems in data entry (may reduce algorithm execution time).
+
+    Returns
+    -------
+    pi : :doc:`numpy array <numpy:reference/generated/numpy.array>`
+        The estimated statistical order (:math:`p_{i}`)
+
+    See Also
+    --------
+    ryan_joiner
+
+
+    Notes
+    -----
+
+    The `alpha` parameter corresponds to the values studied by [1]_, which adopts the following equation to estimate the statistical order:
+
+    .. math::
+
+            p_{i} = \\frac{i - \\alpha}{n - 2 \\times \\alpha + 1}
+
+    where :math:`n` is the sample size and :math:`i` is the ith observation.
+
+    alpha = A is adoptedd in the implementations of the Ryan-Joiner test in Minitab and Statext software. Also, this option is  cited by [2]_ as an alternative.
+
+    `alpha="3/8" is adopted in the implementations of the Ryan-Joiner test in Minitab and Statext software. This option is also cited as an alternative by [2]_.
+
+    References
+    ----------
+    .. [1] BLOM, G. Statistical Estimates and Transformed Beta-Variables. New York: John Wiley and Sons, Inc, p. 71-72, 1958.
+
+    .. [2]  RYAN, T. A., JOINER, B. L. Normal Probability Plots and Tests for Normality, Technical Report, Statistics Department, The Pennsylvania State University, 1976. Available at `www.additive-net.de <https://www.additive-net.de/de/component/jdownloads/send/70-support/236-normal-probability-plots-and-tests-for-normality-thomas-a-ryan-jr-bryan-l-joiner>`_. Access on: 22 Jul. 2023.
+
+
+    Examples
+    --------
+    >>> from normtest import normtest
+    >>> result = normtest.ordered_statistics(7, method="blom")
+    >>> print(result)
+    [0.0862069  0.22413793 0.36206897 0.5        0.63793103 0.77586207
+    0.9137931 ]
+
+    """
+    if safe:
+        parameters.param_options(
+            option=alpha,
+            param_options=["0", "3/8", "1/2"],
+            param_name="alpha",
+            func_name="order_statistic",
+        )
+        types.is_int(
+            value=sample_size, param_name="sample_size", func_name="order_statistic"
+        )
+        numbers.is_greater_than(
+            value=sample_size,
+            lower=4,
+            param_name="sample_size",
+            func_name="order_statistic",
+            inclusive=True,
+        )
+
+    i = np.arange(1, sample_size + 1)
+    if alpha == "3/8":
+        alpha = 3 / 8
+    elif alpha == "0":
+        alpha = 0
+    else:
+        alpha = 0.5
+
+    return (i - alpha) / (sample_size - 2 * alpha + 1)